+++ index.js
@@ -7,4 +7,7 @@ export * from './packages/toast/toast';
 export * from './packages/toast/toast-container';
 export * from './packages/toast/api';
 export * from './packages/broadcast';
+export * from './packages/attention';
+export * from './packages/textfield';
+export * from './packages/affix';
 export * from './packages/utils/expand-transition';